Add support for decoding animated WebP
Decode the first frame of animated WebPs using the demuxer API if
they're found to be animated. This does not add animation support (imlib
doesn't support that) but allows viewing the first frame, which is
much better than nothing.
You would think WebPDecodeBGRA would handle this, but apparently not.

Updates and fixes the makefile as well.